+++ inst/doc/tips-tricks.R
@@ -39,6 +39,16 @@ create_notebook_explore(
 #               output_file = "data_dict_iris.md",
 #               output_dir = tempdir())
 
+## ----message=FALSE, warning=FALSE, fig.width=6, fig.height=4------------------
+data <- use_data_penguins()
+data |> explore(flipper_length_mm, color = "lightskyblue")
+
+## ----message=FALSE, warning=FALSE, fig.width=6, fig.height=4------------------
+data |>
+  drop_obs_with_na() |> 
+  explore(flipper_length_mm, bill_length_mm, 
+          target = sex, color = c("deeppink", "blue"))
+
 ## ----echo=TRUE, message=FALSE, warning=FALSE----------------------------------
 colors <- mix_color("blue", n = 5)
 colors
@@ -51,5 +61,12 @@ colors <- mix_color("gold", "red", n = 4)
 colors
 
 ## ----echo=TRUE, message=FALSE, warning=FALSE----------------------------------
+show_color(colors)
+
+## ----echo=TRUE, message=FALSE, warning=FALSE----------------------------------
+get_color()
+
+## ----echo=TRUE, message=FALSE, warning=FALSE----------------------------------
+colors <- get_color("google")
 show_color(colors)